pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"


[project]
name = "mini_cfg"
authors = [{"name" = "Robert DeMaria", "email" = "robert.demaria@colostate.edu"}]
description = "Small library to assist with storing configuration information in dataclasses."
readme = {file = "README.md", content-type = "text/markdown"}
requires-python = ">=3.11"
dynamic = ["version"]
#license = {file = "LICENSE.txt"}

[project.optional-dependencies]
read_yaml=["pyyaml"]

[project.urls]
Homepage = "https://bear.cira.colostate.edu/rdemaria/mini_cfg"
Documentation = "https://bear.cira.colostate.edu/rdemaria/mini_cfg"
Repository = "https://bear.cira.colostate.edu/rdemaria/mini_cfg.git"

[tool.setuptools.dynamic]
version = {attr = "mini_cfg.__version__"}